rag/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

httpx==0.25.0
python-multipart==0.0.6

# Document processing
Pillow==10.1.0
pymupdf==1.23.8

# OCR (for scanned PDFs)
opencv-python==4.8.1.78
omegaconf==2.3.0
onnxruntime==1.19.2
numpy<2.0.0
pyclipper==1.3.0.post5
shapely==2.0.2

# LangChain (for agent I/O orchestration)
langchain==0.1.0
langchain-openai==0.0.5
langchain-core>=0.1.7,<0.2
langchain-community>=0.0.10

faiss-cpu==1.7.4

# Core ML/AI
# torch>=2.2.0
transformers>=4.37.0